# docs/7_백엔드/openapi_3.0.3.yaml
openapi: 3.0.3
info:
  title: 써도돼요 API
  description: 창업지원금 지출비 사전승인 판정. `프론트 연동 사양.md` §8 계약.
  version: 0.1.0
paths:
...
  /api/l3/upload:
    post:
      tags:
      - L3 업로드
      summary: 업로드
      operationId: 업로드_api_l3_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_____api_l3_upload_post'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/L3_____'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/l3/{doc_id}:
    get:
      tags:
      - L3 업로드
      summary: 상태
      description: 프론트가 「분석 중」 스피너를 돌리며 폴링하는 경로.
      operationId: 상태_api_l3__doc_id__get
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      - name: org_id
        in: query
        required: false
        schema:
          title: Org Id
          type: string
          nullable: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/L3_____'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_____api_l3_upload_post:
      properties:
        파일:
          type: string
          contentMediaType: application/octet-stream
          title: 파일
        org_id:
          type: string
          title: Org Id
        기관명:
          title: 기관명
          type: string
          nullable: true
      type: object
      required:
      - 파일
      - org_id
      title: Body_업로드_api_l3_upload_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    L3_____:
      properties:
        doc_id:
          type: string
          title: Doc Id
        파일명:
          type: string
          title: 파일명
        확장자:
          type: string
          title: 확장자
        상태:
          type: string
          enum:
          - 파싱대기
          - 파싱중
          - 완료
          - 실패
          title: 상태
          default: 파싱대기
        조_건수:
          title: 조 건수
          type: integer
          nullable: true
        dangling:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Dangling
        메시지:
          title: 메시지
          type: string
          nullable: true
      type: object
      required:
      - doc_id
      - 파일명
      - 확장자
      title: L3업로드응답
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...